Storing relative paths for text index directories
